Cleanup: Objective-C: Replace usages of `Nil` with `nil`
In Objective-C, both the `Nil` and `nil` macros can be used to designate
Nullabity. However, only `nil` (with a lower case n) is commonly used,
as can be seen in documents such as the Google Objective-C Style Guide.

In this sense, this commits replaces instances of `Nil` with `nil` for
greater consistency.