Fix #129490: Anim: F-Curve only shown as error when no slot users resolve
Fix a bug where F-Curves in the Action editor were only shown in 'error'
state when none of the slot users could resolve the property.

The error line is now shown whenever the active object (or more
formally, the ID for which the Action editor is drawn) cannot resolve
the F-Curve's RNA path.

The lookup of the RNA path to determine the F-Curve's label is still
looping over all users of the action slot, in a best-effort attempt to
figure out the proper name for the thing. This means that "finding the
name" and "finding the error state" are now at least a little bit
separated from each other.

Pull Request: https://projects.blender.org/blender/blender/pulls/129692