Fix: EEVEE: VR not working
The issue is caused by the new way to compute the projection
matrix that tries to recreate it from view3d properties.
This is needed for proper overscan support.

However, this breaks for VR as the view3d is only partially
setup.

Skip the setup in VR and remove overscan for VR to avoid any
possible issues with it.

This fix avoid changing the behavior of the
`ED_view3d_draw_offscreen` which is exposed from our pyGPU
API. The pyGPU API is kind of broken if using custom
projection matrix + overscan, but that's a different issue.

Fix #125456

Pull Request: https://projects.blender.org/blender/blender/pulls/129813