Libs: Fix static libs not getting linked properly on linux
Multiple of our libraries would pull in system libraries that we have static versions of.
These where libxml2 and libz and also libzstd.

I also noticed that the script that was supposed to check on this was not really usable to check for these kinds of things so I updated it.
Now you can run it with `python3.11 tools/check_blender_release/check_release.py -- ../build_linux_release/bin/` and it will check all binaries and libraries that we ship for any system libraries that we don't expect to be linked.

The libraries I'm aware of that pulled in system libs were:
boost
mesa
osl
llvm

(The removed cmake flags for osl was because these were unused and cmake printed warnings)

Pull Request: https://projects.blender.org/blender/blender/pulls/130236